Acc WCCipher class to handle data encryption
